+++ src/components/ConsoleView/ConsoleView.tsx
@@ -102,7 +102,7 @@ export const ConsoleView = () => {
                     styles={{
                         BASE_FONT_FAMILY: "'DM Mono', monospace",
 
-                        PADDING: "0.4rem 0",
+                        PADDING: "0.425rem 0",
 
                         LOG_BORDER: "oklch(var(--b1))",