+++ src/util/normalizeKey.ts
@@ -1,0 +1,8 @@
+const isMac = navigator.userAgent.includes("Mac");
+
+export default (key: string) => {
+    const k = key.toLocaleLowerCase();
+    return isMac && k.toLowerCase() == "ctrl"
+        ? "cmd"
+        : k.toLocaleLowerCase();
+};